article api ready

+++ controllers/user.controllers.ts
@@ -1,5 +1,6 @@
 import {Request, Response} from 'express';
-import db from '../models'
+import db from '../models';
+import crypto from 'crypto';
 
 const getUsers = (req: Request, res: Response) =>{
     db.User.findAll({ 
@@ -19,8 +20,8 @@ const getUser = (req: Request, res: Response) => {
         userId, 
         { attributes: { exclude: ['password', "updatedAt"]}} 
         )
-    .then((result: any)=>{
-        res.status(200).json(result);
+    .then((user: any)=>{
+        res.status(200).json(user);
     }).catch((err: any)=>{
         res.status(500).json({ error :  err});
     })
@@ -32,9 +33,14 @@ const createUser = (req: Request, res: Response) => {
         email: req.body.email,
         password: req.body.password
     }
-    db.User.create(user).then( (result: any)=>{
-        const toSend = result.dataValues;
+
+    user.password =  crypto.pbkdf2Sync(user.password, "salt", 10000, 100, 'sha512').toString('hex');
+    
+
+    db.User.create(user).then( (user: any)=>{
+        const toSend = user.dataValues;
         delete toSend["password"];
+        toSend["token"] = user.generateJWT();
         res.status(201).json(toSend);
     }).catch( (err: object)=> {
         res.status(500).json({error : err})
@@ -42,31 +48,47 @@ const createUser = (req: Request, res: Response) => {
 }
 
 const updateUser = (req: Request, res: Response) => {
+
+    const jwt = JSON.parse(req.params.jwt);
     const userId = req.params.userId;
-    const user: {[k: string]: any} = {};
-    if(req.body.name) user.name =  req.body.name;
-    if(req.body.email) user.email =  req.body.email;
-    if(req.body.password) user.password =  req.body.password;
-    db.User.update(
-        user,
-        {
-            where: { id: userId }
-        }).then( (result: any)=>{
-        res.status(200).json(result)
-    })
-    .catch( (err: any)=>{
-        res.status(500).json({error : err})
-    })
+    
+    if(userId != jwt.id)
+        res.status(400).json({error : "Incorrect token"})
+    else{
+        const user: {[k: string]: any} = {};
+        if(req.body.name) user.name =  req.body.name;
+        if(req.body.email) user.email =  req.body.email;
+        if(req.body.password) user.password =  req.body.password;
+        
+        db.User.update(
+            user,
+            {
+                where: { id: userId }
+            }).then( (result: any)=>{
+            res.status(200).json(result)
+        })
+        .catch( (err: any)=>{
+            console.log(err);
+            res.status(500).json({error : err})
+        })
+    }
 }
 
 const deleteUser = (req: Request, res: Response) => {
+
+    const jwt = JSON.parse(req.params.jwt);
     const userId = req.params.userId;
-    db.User.destroy( {where: {id : userId} , cascade: true } ).then( (result: any)=>{
-        res.status(200).json(result)
-    })
-    .catch( (err: any)=>{
-        res.status(500).json({error : err})
-    })
+    
+    if(jwt.id != userId){
+        res.status(400).json({error: "Incorrect token"})
+    }
+    else
+        db.User.destroy( {where: {id : userId} , cascade: true } ).then( (result: any)=>{
+            res.status(200).json(result)
+        })
+        .catch( (err: any)=>{
+            res.status(500).json({error : err})
+        })
 }
 
 const getUserArticles = (req: Request, res: Response) =>{
